+++ lib/common/presentation/router/app_router.gr.dart
@@ -9,31 +9,31 @@
 // coverage:ignore-file
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'package:auto_route/auto_route.dart' as _i27;
-import 'package:collection/collection.dart' as _i31;
-import 'package:flutter/material.dart' as _i28;
+import 'package:auto_route/auto_route.dart' as _i28;
+import 'package:collection/collection.dart' as _i32;
+import 'package:flutter/material.dart' as _i29;
 import 'package:zachranobed/app/presentation/home_screen.dart' as _i16;
-import 'package:zachranobed/common/domain/model/user_data.dart' as _i32;
+import 'package:zachranobed/common/domain/model/user_data.dart' as _i33;
 import 'package:zachranobed/features/activepair/presentation/change_active_pair_screen.dart'
     as _i4;
 import 'package:zachranobed/features/appTerms/presentation/app_terms_screen.dart'
     as _i2;
 import 'package:zachranobed/features/debug/components_screen.dart' as _i6;
 import 'package:zachranobed/features/debug/debug_screen.dart' as _i8;
-import 'package:zachranobed/features/faq/domain/model/faq_item.dart' as _i30;
+import 'package:zachranobed/features/faq/domain/model/faq_item.dart' as _i31;
 import 'package:zachranobed/features/faq/presentation/screen/faq_detail_screen.dart'
     as _i10;
 import 'package:zachranobed/features/faq/presentation/screen/faq_questions_screen.dart'
     as _i11;
 import 'package:zachranobed/features/food/domain/model/box_delivery_confirmation.dart'
-    as _i29;
+    as _i30;
 import 'package:zachranobed/features/food/domain/model/food_box_statistics.dart'
-    as _i33;
-import 'package:zachranobed/features/food/domain/model/food_box_type.dart'
-    as _i35;
-import 'package:zachranobed/features/food/domain/model/food_info.dart' as _i36;
-import 'package:zachranobed/features/food/domain/model/meal_suggestion.dart'
     as _i34;
+import 'package:zachranobed/features/food/domain/model/food_box_type.dart'
+    as _i36;
+import 'package:zachranobed/features/food/domain/model/food_info.dart' as _i37;
+import 'package:zachranobed/features/food/domain/model/meal_suggestion.dart'
+    as _i35;
 import 'package:zachranobed/features/food/presentation/screens/add_meals_to_history_screen.dart'
     as _i1;
 import 'package:zachranobed/features/food/presentation/screens/box_delivery_confirmation_screen.dart'
@@ -45,43 +45,45 @@ import 'package:zachranobed/features/food/presentation/screens/food_boxes_checku
 import 'package:zachranobed/features/food/presentation/screens/food_boxes_detail_screen.dart'
     as _i13;
 import 'package:zachranobed/features/food/presentation/screens/meal_suggestion_form_screen.dart'
-    as _i18;
-import 'package:zachranobed/features/food/presentation/screens/meal_suggestions_list_screen.dart'
     as _i19;
-import 'package:zachranobed/features/food/presentation/screens/offer_food_boxes_screen.dart'
-    as _i21;
-import 'package:zachranobed/features/food/presentation/screens/offer_food_detail_screen.dart'
+import 'package:zachranobed/features/food/presentation/screens/meal_suggestions_list_screen.dart'
     as _i20;
-import 'package:zachranobed/features/food/presentation/screens/offer_food_overview_screen.dart'
+import 'package:zachranobed/features/food/presentation/screens/offer_food_boxes_screen.dart'
     as _i22;
-import 'package:zachranobed/features/food/presentation/screens/order_shipping_of_boxes_screen.dart'
+import 'package:zachranobed/features/food/presentation/screens/offer_food_detail_screen.dart'
+    as _i21;
+import 'package:zachranobed/features/food/presentation/screens/offer_food_overview_screen.dart'
     as _i23;
+import 'package:zachranobed/features/food/presentation/screens/order_shipping_of_boxes_screen.dart'
+    as _i24;
 import 'package:zachranobed/features/food/presentation/screens/thank_you_screen.dart'
-    as _i25;
+    as _i26;
 import 'package:zachranobed/features/forceupdate/presentation/force_update_screen.dart'
     as _i14;
+import 'package:zachranobed/features/inactiveaccount/presentation/inactive_account_screen.dart'
+    as _i17;
 import 'package:zachranobed/features/login/presentation/screen/change_password_screen.dart'
     as _i5;
 import 'package:zachranobed/features/login/presentation/screen/forgot_password_screen.dart'
     as _i15;
 import 'package:zachranobed/features/login/presentation/screen/login_screen.dart'
-    as _i17;
+    as _i18;
 import 'package:zachranobed/features/menu/presentation/contacts_screen.dart'
     as _i7;
 import 'package:zachranobed/features/menu/presentation/profile_screen.dart'
-    as _i24;
+    as _i25;
 import 'package:zachranobed/features/whatsnew/presentation/whats_new_screen.dart'
-    as _i26;
+    as _i27;
 
 /// generated route for
 /// [_i1.AddMealsToHistoryScreen]
-class AddMealsToHistoryRoute extends _i27.PageRouteInfo<void> {
-  const AddMealsToHistoryRoute({List<_i27.PageRouteInfo>? children})
+class AddMealsToHistoryRoute extends _i28.PageRouteInfo<void> {
+  const AddMealsToHistoryRoute({List<_i28.PageRouteInfo>? children})
       : super(AddMealsToHistoryRoute.name, initialChildren: children);
 
   static const String name = 'AddMealsToHistoryRoute';
 
-  static _i27.PageInfo page = _i27.PageInfo(
+  static _i28.PageInfo page = _i28.PageInfo(
     name,
     builder: (data) {
       return const _i1.AddMealsToHistoryScreen();
@@ -91,11 +93,11 @@ class AddMealsToHistoryRoute extends _i27.PageRouteInfo<void> {
 
 /// generated route for
 /// [_i2.AppTermsScreen]
-class AppTermsRoute extends _i27.PageRouteInfo<AppTermsRouteArgs> {
+class AppTermsRoute extends _i28.PageRouteInfo<AppTermsRouteArgs> {
   AppTermsRoute({
-    _i28.Key? key,
+    _i29.Key? key,
     required bool hasNoAcceptedVersion,
-    List<_i27.PageRouteInfo>? children,
+    List<_i28.PageRouteInfo>? children,
   }) : super(
           AppTermsRoute.name,
           args: AppTermsRouteArgs(
@@ -107,7 +109,7 @@ class AppTermsRoute extends _i27.PageRouteInfo<AppTermsRouteArgs> {
 
   static const String name = 'AppTermsRoute';
 
-  static _i27.PageInfo page = _i27.PageInfo(
+  static _i28.PageInfo page = _i28.PageInfo(
     name,
     builder: (data) {
       final args = data.argsAs<AppTermsRouteArgs>();
@@ -122,7 +124,7 @@ class AppTermsRoute extends _i27.PageRouteInfo<AppTermsRouteArgs> {
 class AppTermsRouteArgs {
   const AppTermsRouteArgs({this.key, required this.hasNoAcceptedVersion});
 
-  final _i28.Key? key;
+  final _i29.Key? key;
 
   final bool hasNoAcceptedVersion;
 
@@ -146,11 +148,11 @@ class AppTermsRouteArgs {
 /// generated route for
 /// [_i3.BoxDeliveryConfirmationScreen]
 class BoxDeliveryConfirmationRoute
-    extends _i27.PageRouteInfo<BoxDeliveryConfirmationRouteArgs> {
+    extends _i28.PageRouteInfo<BoxDeliveryConfirmationRouteArgs> {
   BoxDeliveryConfirmationRoute({
-    _i28.Key? key,
-    required _i29.BoxDeliveryConfirmation confirmation,
-    List<_i27.PageRouteInfo>? children,
+    _i29.Key? key,
+    required _i30.BoxDeliveryConfirmation confirmation,
+    List<_i28.PageRouteInfo>? children,
   }) : super(
           BoxDeliveryConfirmationRoute.name,
           args: BoxDeliveryConfirmationRouteArgs(
@@ -162,7 +164,7 @@ class BoxDeliveryConfirmationRoute
 
   static const String name = 'BoxDeliveryConfirmationRoute';
 
-  static _i27.PageInfo page = _i27.PageInfo(
+  static _i28.PageInfo page = _i28.PageInfo(
     name,
     builder: (data) {
       final args = data.argsAs<BoxDeliveryConfirmationRouteArgs>();
@@ -180,9 +182,9 @@ class BoxDeliveryConfirmationRouteArgs {
     required this.confirmation,
   });
 
-  final _i28.Key? key;
+  final _i29.Key? key;
 
-  final _i29.BoxDeliveryConfirmation confirmation;
+  final _i30.BoxDeliveryConfirmation confirmation;
 
   @override
   String toString() {
@@ -202,13 +204,13 @@ class BoxDeliveryConfirmationRouteArgs {
 
 /// generated route for
 /// [_i4.ChangeActivePairScreen]
-class ChangeActivePairRoute extends _i27.PageRouteInfo<void> {
-  const ChangeActivePairRoute({List<_i27.PageRouteInfo>? children})
+class ChangeActivePairRoute extends _i28.PageRouteInfo<void> {
+  const ChangeActivePairRoute({List<_i28.PageRouteInfo>? children})
       : super(ChangeActivePairRoute.name, initialChildren: children);
 
   static const String name = 'ChangeActivePairRoute';
 
-  static _i27.PageInfo page = _i27.PageInfo(
+  static _i28.PageInfo page = _i28.PageInfo(
     name,
     builder: (data) {
       return const _i4.ChangeActivePairScreen();
@@ -218,13 +220,13 @@ class ChangeActivePairRoute extends _i27.PageRouteInfo<void> {
 
 /// generated route for
 /// [_i5.ChangePasswordScreen]
-class ChangePasswordRoute extends _i27.PageRouteInfo<void> {
-  const ChangePasswordRoute({List<_i27.PageRouteInfo>? children})
+class ChangePasswordRoute extends _i28.PageRouteInfo<void> {
+  const ChangePasswordRoute({List<_i28.PageRouteInfo>? children})
       : super(ChangePasswordRoute.name, initialChildren: children);
 
   static const String name = 'ChangePasswordRoute';
 
-  static _i27.PageInfo page = _i27.PageInfo(
+  static _i28.PageInfo page = _i28.PageInfo(
     name,
     builder: (data) {
       return const _i5.ChangePasswordScreen();
@@ -234,13 +236,13 @@ class ChangePasswordRoute extends _i27.PageRouteInfo<void> {
 
 /// generated route for
 /// [_i6.ComponentsScreen]
-class ComponentsRoute extends _i27.PageRouteInfo<void> {
-  const ComponentsRoute({List<_i27.PageRouteInfo>? children})
+class ComponentsRoute extends _i28.PageRouteInfo<void> {
+  const ComponentsRoute({List<_i28.PageRouteInfo>? children})
       : super(ComponentsRoute.name, initialChildren: children);
 
   static const String name = 'ComponentsRoute';
 
-  static _i27.PageInfo page = _i27.PageInfo(
+  static _i28.PageInfo page = _i28.PageInfo(
     name,
     builder: (data) {
       return const _i6.ComponentsScreen();
@@ -250,13 +252,13 @@ class ComponentsRoute extends _i27.PageRouteInfo<void> {
 
 /// generated route for
 /// [_i7.ContactsScreen]
-class ContactsRoute extends _i27.PageRouteInfo<void> {
-  const ContactsRoute({List<_i27.PageRouteInfo>? children})
+class ContactsRoute extends _i28.PageRouteInfo<void> {
+  const ContactsRoute({List<_i28.PageRouteInfo>? children})
       : super(ContactsRoute.name, initialChildren: children);
 
   static const String name = 'ContactsRoute';
 
-  static _i27.PageInfo page = _i27.PageInfo(
+  static _i28.PageInfo page = _i28.PageInfo(
     name,
     builder: (data) {
       return const _i7.ContactsScreen();
@@ -266,13 +268,13 @@ class ContactsRoute extends _i27.PageRouteInfo<void> {
 
 /// generated route for
 /// [_i8.DebugScreen]
-class DebugRoute extends _i27.PageRouteInfo<void> {
-  const DebugRoute({List<_i27.PageRouteInfo>? children})
+class DebugRoute extends _i28.PageRouteInfo<void> {
+  const DebugRoute({List<_i28.PageRouteInfo>? children})
       : super(DebugRoute.name, initialChildren: children);
 
   static const String name = 'DebugRoute';
 
-  static _i27.PageInfo page = _i27.PageInfo(
+  static _i28.PageInfo page = _i28.PageInfo(
     name,
     builder: (data) {
       return const _i8.DebugScreen();
@@ -282,11 +284,11 @@ class DebugRoute extends _i27.PageRouteInfo<void> {
 
 /// generated route for
 /// [_i9.DeliveryDetailScreen]
-class DeliveryDetailRoute extends _i27.PageRouteInfo<DeliveryDetailRouteArgs> {
+class DeliveryDetailRoute extends _i28.PageRouteInfo<DeliveryDetailRouteArgs> {
   DeliveryDetailRoute({
-    _i28.Key? key,
+    _i29.Key? key,
     required String deliveryId,
-    List<_i27.PageRouteInfo>? children,
+    List<_i28.PageRouteInfo>? children,
   }) : super(
           DeliveryDetailRoute.name,
           args: DeliveryDetailRouteArgs(key: key, deliveryId: deliveryId),
@@ -295,7 +297,7 @@ class DeliveryDetailRoute extends _i27.PageRouteInfo<DeliveryDetailRouteArgs> {
 
   static const String name = 'DeliveryDetailRoute';
 
-  static _i27.PageInfo page = _i27.PageInfo(
+  static _i28.PageInfo page = _i28.PageInfo(
     name,
     builder: (data) {
       final args = data.argsAs<DeliveryDetailRouteArgs>();
@@ -310,7 +312,7 @@ class DeliveryDetailRoute extends _i27.PageRouteInfo<DeliveryDetailRouteArgs> {
 class DeliveryDetailRouteArgs {
   const DeliveryDetailRouteArgs({this.key, required this.deliveryId});
 
-  final _i28.Key? key;
+  final _i29.Key? key;
 
   final String deliveryId;
 
@@ -332,11 +334,11 @@ class DeliveryDetailRouteArgs {
 
 /// generated route for
 /// [_i10.FaqDetailScreen]
-class FaqDetailRoute extends _i27.PageRouteInfo<FaqDetailRouteArgs> {
+class FaqDetailRoute extends _i28.PageRouteInfo<FaqDetailRouteArgs> {
   FaqDetailRoute({
-    _i28.Key? key,
-    required _i30.FaqItem item,
-    List<_i27.PageRouteInfo>? children,
+    _i29.Key? key,
+    required _i31.FaqItem item,
+    List<_i28.PageRouteInfo>? children,
   }) : super(
           FaqDetailRoute.name,
           args: FaqDetailRouteArgs(key: key, item: item),
@@ -345,7 +347,7 @@ class FaqDetailRoute extends _i27.PageRouteInfo<FaqDetailRouteArgs> {
 
   static const String name = 'FaqDetailRoute';
 
-  static _i27.PageInfo page = _i27.PageInfo(
+  static _i28.PageInfo page = _i28.PageInfo(
     name,
     builder: (data) {
       final args = data.argsAs<FaqDetailRouteArgs>();
@@ -357,9 +359,9 @@ class FaqDetailRoute extends _i27.PageRouteInfo<FaqDetailRouteArgs> {
 class FaqDetailRouteArgs {
   const FaqDetailRouteArgs({this.key, required this.item});
 
-  final _i28.Key? key;
+  final _i29.Key? key;
 
-  final _i30.FaqItem item;
+  final _i31.FaqItem item;
 
   @override
   String toString() {
@@ -379,11 +381,11 @@ class FaqDetailRouteArgs {
 
 /// generated route for
 /// [_i11.FaqQuestionsScreen]
-class FaqQuestionsRoute extends _i27.PageRouteInfo<FaqQuestionsRouteArgs> {
+class FaqQuestionsRoute extends _i28.PageRouteInfo<FaqQuestionsRouteArgs> {
   FaqQuestionsRoute({
-    _i28.Key? key,
-    required List<_i30.FaqItem> items,
-    List<_i27.PageRouteInfo>? children,
+    _i29.Key? key,
+    required List<_i31.FaqItem> items,
+    List<_i28.PageRouteInfo>? children,
   }) : super(
           FaqQuestionsRoute.name,
           args: FaqQuestionsRouteArgs(key: key, items: items),
@@ -392,7 +394,7 @@ class FaqQuestionsRoute extends _i27.PageRouteInfo<FaqQuestionsRouteArgs> {
 
   static const String name = 'FaqQuestionsRoute';
 
-  static _i27.PageInfo page = _i27.PageInfo(
+  static _i28.PageInfo page = _i28.PageInfo(
     name,
     builder: (data) {
       final args = data.argsAs<FaqQuestionsRouteArgs>();
@@ -404,9 +406,9 @@ class FaqQuestionsRoute extends _i27.PageRouteInfo<FaqQuestionsRouteArgs> {
 class FaqQuestionsRouteArgs {
   const FaqQuestionsRouteArgs({this.key, required this.items});
 
-  final _i28.Key? key;
+  final _i29.Key? key;
 
-  final List<_i30.FaqItem> items;
+  final List<_i31.FaqItem> items;
 
   @override
   String toString() {
@@ -418,22 +420,22 @@ class FaqQuestionsRouteArgs {
     if (identical(this, other)) return true;
     if (other is! FaqQuestionsRouteArgs) return false;
     return key == other.key &&
-        const _i31.ListEquality().equals(items, other.items);
+        const _i32.ListEquality().equals(items, other.items);
   }
 
   @override
-  int get hashCode => key.hashCode ^ const _i31.ListEquality().hash(items);
+  int get hashCode => key.hashCode ^ const _i32.ListEquality().hash(items);
 }
 
 /// generated route for
 /// [_i12.FoodBoxesCheckupMismatchScreen]
 class FoodBoxesCheckupMismatchRoute
-    extends _i27.PageRouteInfo<FoodBoxesCheckupMismatchRouteArgs> {
+    extends _i28.PageRouteInfo<FoodBoxesCheckupMismatchRouteArgs> {
   FoodBoxesCheckupMismatchRoute({
-    _i28.Key? key,
-    required _i32.UserData user,
-    required List<_i33.FoodBoxStatistics> statistics,
-    List<_i27.PageRouteInfo>? children,
+    _i29.Key? key,
+    required _i33.UserData user,
+    required List<_i34.FoodBoxStatistics> statistics,
+    List<_i28.PageRouteInfo>? children,
   }) : super(
           FoodBoxesCheckupMismatchRoute.name,
           args: FoodBoxesCheckupMismatchRouteArgs(
@@ -446,7 +448,7 @@ class FoodBoxesCheckupMismatchRoute
 
   static const String name = 'FoodBoxesCheckupMismatchRoute';
 
-  static _i27.PageInfo page = _i27.PageInfo(
+  static _i28.PageInfo page = _i28.PageInfo(
     name,
     builder: (data) {
       final args = data.argsAs<FoodBoxesCheckupMismatchRouteArgs>();
@@ -466,11 +468,11 @@ class FoodBoxesCheckupMismatchRouteArgs {
     required this.statistics,
   });
 
-  final _i28.Key? key;
+  final _i29.Key? key;
 
-  final _i32.UserData user;
+  final _i33.UserData user;
 
-  final List<_i33.FoodBoxStatistics> statistics;
+  final List<_i34.FoodBoxStatistics> statistics;
 
   @override
   String toString() {
@@ -483,23 +485,23 @@ class FoodBoxesCheckupMismatchRouteArgs {
     if (other is! FoodBoxesCheckupMismatchRouteArgs) return false;
     return key == other.key &&
         user == other.user &&
-        const _i31.ListEquality().equals(statistics, other.statistics);
+        const _i32.ListEquality().equals(statistics, other.statistics);
   }
 
   @override
   int get hashCode =>
-      key.hashCode ^ user.hashCode ^ const _i31.ListEquality().hash(statistics);
+      key.hashCode ^ user.hashCode ^ const _i32.ListEquality().hash(statistics);
 }
 
 /// generated route for
 /// [_i13.FoodBoxesDetailScreen]
 class FoodBoxesDetailRoute
-    extends _i27.PageRouteInfo<FoodBoxesDetailRouteArgs> {
+    extends _i28.PageRouteInfo<FoodBoxesDetailRouteArgs> {
   FoodBoxesDetailRoute({
-    _i28.Key? key,
-    required _i32.UserData user,
+    _i29.Key? key,
+    required _i33.UserData user,
     bool isCheckupMode = false,
-    List<_i27.PageRouteInfo>? children,
+    List<_i28.PageRouteInfo>? children,
   }) : super(
           FoodBoxesDetailRoute.name,
           args: FoodBoxesDetailRouteArgs(
@@ -512,7 +514,7 @@ class FoodBoxesDetailRoute
 
   static const String name = 'FoodBoxesDetailRoute';
 
-  static _i27.PageInfo page = _i27.PageInfo(
+  static _i28.PageInfo page = _i28.PageInfo(
     name,
     builder: (data) {
       final args = data.argsAs<FoodBoxesDetailRouteArgs>();
@@ -532,9 +534,9 @@ class FoodBoxesDetailRouteArgs {
     this.isCheckupMode = false,
   });
 
-  final _i28.Key? key;
+  final _i29.Key? key;
 
-  final _i32.UserData user;
+  final _i33.UserData user;
 
   final bool isCheckupMode;
 
@@ -558,13 +560,13 @@ class FoodBoxesDetailRouteArgs {
 
 /// generated route for
 /// [_i14.ForceUpdateScreen]
-class ForceUpdateRoute extends _i27.PageRouteInfo<void> {
-  const ForceUpdateRoute({List<_i27.PageRouteInfo>? children})
+class ForceUpdateRoute extends _i28.PageRouteInfo<void> {
+  const ForceUpdateRoute({List<_i28.PageRouteInfo>? children})
       : super(ForceUpdateRoute.name, initialChildren: children);
 
   static const String name = 'ForceUpdateRoute';
 
-  static _i27.PageInfo page = _i27.PageInfo(
+  static _i28.PageInfo page = _i28.PageInfo(
     name,
     builder: (data) {
       return const _i14.ForceUpdateScreen();
@@ -574,13 +576,13 @@ class ForceUpdateRoute extends _i27.PageRouteInfo<void> {
 
 /// generated route for
 /// [_i15.ForgotPasswordScreen]
-class ForgotPasswordRoute extends _i27.PageRouteInfo<void> {
-  const ForgotPasswordRoute({List<_i27.PageRouteInfo>? children})
+class ForgotPasswordRoute extends _i28.PageRouteInfo<void> {
+  const ForgotPasswordRoute({List<_i28.PageRouteInfo>? children})
       : super(ForgotPasswordRoute.name, initialChildren: children);
 
   static const String name = 'ForgotPasswordRoute';
 
-  static _i27.PageInfo page = _i27.PageInfo(
+  static _i28.PageInfo page = _i28.PageInfo(
     name,
     builder: (data) {
       return const _i15.ForgotPasswordScreen();
@@ -590,11 +592,11 @@ class ForgotPasswordRoute extends _i27.PageRouteInfo<void> {
 
 /// generated route for
 /// [_i16.HomeScreen]
-class HomeRoute extends _i27.PageRouteInfo<HomeRouteArgs> {
+class HomeRoute extends _i28.PageRouteInfo<HomeRouteArgs> {
   HomeRoute({
-    _i28.Key? key,
+    _i29.Key? key,
     int initialTabIndex = 0,
-    List<_i27.PageRouteInfo>? children,
+    List<_i28.PageRouteInfo>? children,
   }) : super(
           HomeRoute.name,
           args: HomeRouteArgs(key: key, initialTabIndex: initialTabIndex),
@@ -603,7 +605,7 @@ class HomeRoute extends _i27.PageRouteInfo<HomeRouteArgs> {
 
   static const String name = 'HomeRoute';
 
-  static _i27.PageInfo page = _i27.PageInfo(
+  static _i28.PageInfo page = _i28.PageInfo(
     name,
     builder: (data) {
       final args = data.argsAs<HomeRouteArgs>(
@@ -620,7 +622,7 @@ class HomeRoute extends _i27.PageRouteInfo<HomeRouteArgs> {
 class HomeRouteArgs {
   const HomeRouteArgs({this.key, this.initialTabIndex = 0});
 
-  final _i28.Key? key;
+  final _i29.Key? key;
 
   final int initialTabIndex;
 
@@ -641,45 +643,95 @@ class HomeRouteArgs {
 }
 
 /// generated route for
-/// [_i17.LoginScreen]
-class LoginRoute extends _i27.PageRouteInfo<void> {
-  const LoginRoute({List<_i27.PageRouteInfo>? children})
+/// [_i17.InactiveAccountScreen]
+class InactiveAccountRoute
+    extends _i28.PageRouteInfo<InactiveAccountRouteArgs> {
+  InactiveAccountRoute({
+    _i29.Key? key,
+    String? entityId,
+    List<_i28.PageRouteInfo>? children,
+  }) : super(
+          InactiveAccountRoute.name,
+          args: InactiveAccountRouteArgs(key: key, entityId: entityId),
+          initialChildren: children,
+        );
+
+  static const String name = 'InactiveAccountRoute';
+
+  static _i28.PageInfo page = _i28.PageInfo(
+    name,
+    builder: (data) {
+      final args = data.argsAs<InactiveAccountRouteArgs>(
+        orElse: () => const InactiveAccountRouteArgs(),
+      );
+      return _i17.InactiveAccountScreen(key: args.key, entityId: args.entityId);
+    },
+  );
+}
+
+class InactiveAccountRouteArgs {
+  const InactiveAccountRouteArgs({this.key, this.entityId});
+
+  final _i29.Key? key;
+
+  final String? entityId;
+
+  @override
+  String toString() {
+    return 'InactiveAccountRouteArgs{key: $key, entityId: $entityId}';
+  }
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(this, other)) return true;
+    if (other is! InactiveAccountRouteArgs) return false;
+    return key == other.key && entityId == other.entityId;
+  }
+
+  @override
+  int get hashCode => key.hashCode ^ entityId.hashCode;
+}
+
+/// generated route for
+/// [_i18.LoginScreen]
+class LoginRoute extends _i28.PageRouteInfo<void> {
+  const LoginRoute({List<_i28.PageRouteInfo>? children})
       : super(LoginRoute.name, initialChildren: children);
 
   static const String name = 'LoginRoute';
 
-  static _i27.PageInfo page = _i27.PageInfo(
+  static _i28.PageInfo page = _i28.PageInfo(
     name,
     builder: (data) {
-      return const _i17.LoginScreen();
+      return const _i18.LoginScreen();
     },
   );
 }
 
 /// generated route for
-/// [_i18.MealSuggestionAddScreen]
-class MealSuggestionAddRoute extends _i27.PageRouteInfo<void> {
-  const MealSuggestionAddRoute({List<_i27.PageRouteInfo>? children})
+/// [_i19.MealSuggestionAddScreen]
+class MealSuggestionAddRoute extends _i28.PageRouteInfo<void> {
+  const MealSuggestionAddRoute({List<_i28.PageRouteInfo>? children})
       : super(MealSuggestionAddRoute.name, initialChildren: children);
 
   static const String name = 'MealSuggestionAddRoute';
 
-  static _i27.PageInfo page = _i27.PageInfo(
+  static _i28.PageInfo page = _i28.PageInfo(
     name,
     builder: (data) {
-      return const _i18.MealSuggestionAddScreen();
+      return const _i19.MealSuggestionAddScreen();
     },
   );
 }
 
 /// generated route for
-/// [_i18.MealSuggestionEditScreen]
+/// [_i19.MealSuggestionEditScreen]
 class MealSuggestionEditRoute
-    extends _i27.PageRouteInfo<MealSuggestionEditRouteArgs> {
+    extends _i28.PageRouteInfo<MealSuggestionEditRouteArgs> {
   MealSuggestionEditRoute({
-    _i28.Key? key,
-    required _i34.MealSuggestion suggestion,
-    List<_i27.PageRouteInfo>? children,
+    _i29.Key? key,
+    required _i35.MealSuggestion suggestion,
+    List<_i28.PageRouteInfo>? children,
   }) : super(
           MealSuggestionEditRoute.name,
           args: MealSuggestionEditRouteArgs(key: key, suggestion: suggestion),
@@ -688,11 +740,11 @@ class MealSuggestionEditRoute
 
   static const String name = 'MealSuggestionEditRoute';
 
-  static _i27.PageInfo page = _i27.PageInfo(
+  static _i28.PageInfo page = _i28.PageInfo(
     name,
     builder: (data) {
       final args = data.argsAs<MealSuggestionEditRouteArgs>();
-      return _i18.MealSuggestionEditScreen(
+      return _i19.MealSuggestionEditScreen(
         key: args.key,
         suggestion: args.suggestion,
       );
@@ -703,9 +755,9 @@ class MealSuggestionEditRoute
 class MealSuggestionEditRouteArgs {
   const MealSuggestionEditRouteArgs({this.key, required this.suggestion});
 
-  final _i28.Key? key;
+  final _i29.Key? key;
 
-  final _i34.MealSuggestion suggestion;
+  final _i35.MealSuggestion suggestion;
 
   @override
   String toString() {
@@ -724,44 +776,44 @@ class MealSuggestionEditRouteArgs {
 }
 
 /// generated route for
-/// [_i19.MealSuggestionsListScreen]
-class MealSuggestionsListRoute extends _i27.PageRouteInfo<void> {
-  const MealSuggestionsListRoute({List<_i27.PageRouteInfo>? children})
+/// [_i20.MealSuggestionsListScreen]
+class MealSuggestionsListRoute extends _i28.PageRouteInfo<void> {
+  const MealSuggestionsListRoute({List<_i28.PageRouteInfo>? children})
       : super(MealSuggestionsListRoute.name, initialChildren: children);
 
   static const String name = 'MealSuggestionsListRoute';
 
-  static _i27.PageInfo page = _i27.PageInfo(
+  static _i28.PageInfo page = _i28.PageInfo(
     name,
     builder: (data) {
-      return const _i19.MealSuggestionsListScreen();
+      return const _i20.MealSuggestionsListScreen();
     },
   );
 }
 
 /// generated route for
-/// [_i20.OfferFoodAddNewScreen]
-class OfferFoodAddNewRoute extends _i27.PageRouteInfo<void> {
-  const OfferFoodAddNewRoute({List<_i27.PageRouteInfo>? children})
+/// [_i21.OfferFoodAddNewScreen]
+class OfferFoodAddNewRoute extends _i28.PageRouteInfo<void> {
+  const OfferFoodAddNewRoute({List<_i28.PageRouteInfo>? children})
       : super(OfferFoodAddNewRoute.name, initialChildren: children);
 
   static const String name = 'OfferFoodAddNewRoute';
 
-  static _i27.PageInfo page = _i27.PageInfo(
+  static _i28.PageInfo page = _i28.PageInfo(
     name,
     builder: (data) {
-      return const _i20.OfferFoodAddNewScreen();
+      return const _i21.OfferFoodAddNewScreen();
     },
   );
 }
 
 /// generated route for
-/// [_i21.OfferFoodBoxesScreen]
-class OfferFoodBoxesRoute extends _i27.PageRouteInfo<OfferFoodBoxesRouteArgs> {
+/// [_i22.OfferFoodBoxesScreen]
+class OfferFoodBoxesRoute extends _i28.PageRouteInfo<OfferFoodBoxesRouteArgs> {
   OfferFoodBoxesRoute({
-    _i28.Key? key,
-    required Map<_i35.FoodBoxType, int> currentBoxesQuantity,
-    List<_i27.PageRouteInfo>? children,
+    _i29.Key? key,
+    required Map<_i36.FoodBoxType, int> currentBoxesQuantity,
+    List<_i28.PageRouteInfo>? children,
   }) : super(
           OfferFoodBoxesRoute.name,
           args: OfferFoodBoxesRouteArgs(
@@ -773,11 +825,11 @@ class OfferFoodBoxesRoute extends _i27.PageRouteInfo<OfferFoodBoxesRouteArgs> {
 
   static const String name = 'OfferFoodBoxesRoute';
 
-  static _i27.PageInfo page = _i27.PageInfo(
+  static _i28.PageInfo page = _i28.PageInfo(
     name,
     builder: (data) {
       final args = data.argsAs<OfferFoodBoxesRouteArgs>();
-      return _i21.OfferFoodBoxesScreen(
+      return _i22.OfferFoodBoxesScreen(
         key: args.key,
         currentBoxesQuantity: args.currentBoxesQuantity,
       );
@@ -788,9 +840,9 @@ class OfferFoodBoxesRoute extends _i27.PageRouteInfo<OfferFoodBoxesRouteArgs> {
 class OfferFoodBoxesRouteArgs {
   const OfferFoodBoxesRouteArgs({this.key, required this.currentBoxesQuantity});
 
-  final _i28.Key? key;
+  final _i29.Key? key;
 
-  final Map<_i35.FoodBoxType, int> currentBoxesQuantity;
+  final Map<_i36.FoodBoxType, int> currentBoxesQuantity;
 
   @override
   String toString() {
@@ -802,7 +854,7 @@ class OfferFoodBoxesRouteArgs {
     if (identical(this, other)) return true;
     if (other is! OfferFoodBoxesRouteArgs) return false;
     return key == other.key &&
-        const _i31.MapEquality().equals(
+        const _i32.MapEquality().equals(
           currentBoxesQuantity,
           other.currentBoxesQuantity,
         );
@@ -810,17 +862,17 @@ class OfferFoodBoxesRouteArgs {
 
   @override
   int get hashCode =>
-      key.hashCode ^ const _i31.MapEquality().hash(currentBoxesQuantity);
+      key.hashCode ^ const _i32.MapEquality().hash(currentBoxesQuantity);
 }
 
 /// generated route for
-/// [_i20.OfferFoodEditExistingScreen]
+/// [_i21.OfferFoodEditExistingScreen]
 class OfferFoodEditExistingRoute
-    extends _i27.PageRouteInfo<OfferFoodEditExistingRouteArgs> {
+    extends _i28.PageRouteInfo<OfferFoodEditExistingRouteArgs> {
   OfferFoodEditExistingRoute({
-    _i28.Key? key,
-    required _i36.FoodInfo foodInfo,
-    List<_i27.PageRouteInfo>? children,
+    _i29.Key? key,
+    required _i37.FoodInfo foodInfo,
+    List<_i28.PageRouteInfo>? children,
   }) : super(
           OfferFoodEditExistingRoute.name,
           args: OfferFoodEditExistingRouteArgs(key: key, foodInfo: foodInfo),
@@ -829,11 +881,11 @@ class OfferFoodEditExistingRoute
 
   static const String name = 'OfferFoodEditExistingRoute';
 
-  static _i27.PageInfo page = _i27.PageInfo(
+  static _i28.PageInfo page = _i28.PageInfo(
     name,
     builder: (data) {
       final args = data.argsAs<OfferFoodEditExistingRouteArgs>();
-      return _i20.OfferFoodEditExistingScreen(
+      return _i21.OfferFoodEditExistingScreen(
         key: args.key,
         foodInfo: args.foodInfo,
       );
@@ -844,9 +896,9 @@ class OfferFoodEditExistingRoute
 class OfferFoodEditExistingRouteArgs {
   const OfferFoodEditExistingRouteArgs({this.key, required this.foodInfo});
 
-  final _i28.Key? key;
+  final _i29.Key? key;
 
-  final _i36.FoodInfo foodInfo;
+  final _i37.FoodInfo foodInfo;
 
   @override
   String toString() {
@@ -865,29 +917,29 @@ class OfferFoodEditExistingRouteArgs {
 }
 
 /// generated route for
-/// [_i20.OfferFoodInitialScreen]
-class OfferFoodInitialRoute extends _i27.PageRouteInfo<void> {
-  const OfferFoodInitialRoute({List<_i27.PageRouteInfo>? children})
+/// [_i21.OfferFoodInitialScreen]
+class OfferFoodInitialRoute extends _i28.PageRouteInfo<void> {
+  const OfferFoodInitialRoute({List<_i28.PageRouteInfo>? children})
       : super(OfferFoodInitialRoute.name, initialChildren: children);
 
   static const String name = 'OfferFoodInitialRoute';
 
-  static _i27.PageInfo page = _i27.PageInfo(
+  static _i28.PageInfo page = _i28.PageInfo(
     name,
     builder: (data) {
-      return const _i20.OfferFoodInitialScreen();
+      return const _i21.OfferFoodInitialScreen();
     },
   );
 }
 
 /// generated route for
-/// [_i22.OfferFoodOverviewScreen]
+/// [_i23.OfferFoodOverviewScreen]
 class OfferFoodOverviewRoute
-    extends _i27.PageRouteInfo<OfferFoodOverviewRouteArgs> {
+    extends _i28.PageRouteInfo<OfferFoodOverviewRouteArgs> {
   OfferFoodOverviewRoute({
-    _i28.Key? key,
-    required List<_i36.FoodInfo> initialFoodInfos,
-    List<_i27.PageRouteInfo>? children,
+    _i29.Key? key,
+    required List<_i37.FoodInfo> initialFoodInfos,
+    List<_i28.PageRouteInfo>? children,
   }) : super(
           OfferFoodOverviewRoute.name,
           args: OfferFoodOverviewRouteArgs(
@@ -899,11 +951,11 @@ class OfferFoodOverviewRoute
 
   static const String name = 'OfferFoodOverviewRoute';
 
-  static _i27.PageInfo page = _i27.PageInfo(
+  static _i28.PageInfo page = _i28.PageInfo(
     name,
     builder: (data) {
       final args = data.argsAs<OfferFoodOverviewRouteArgs>();
-      return _i22.OfferFoodOverviewScreen(
+      return _i23.OfferFoodOverviewScreen(
         key: args.key,
         initialFoodInfos: args.initialFoodInfos,
       );
@@ -914,9 +966,9 @@ class OfferFoodOverviewRoute
 class OfferFoodOverviewRouteArgs {
   const OfferFoodOverviewRouteArgs({this.key, required this.initialFoodInfos});
 
-  final _i28.Key? key;
+  final _i29.Key? key;
 
-  final List<_i36.FoodInfo> initialFoodInfos;
+  final List<_i37.FoodInfo> initialFoodInfos;
 
   @override
   String toString() {
@@ -928,7 +980,7 @@ class OfferFoodOverviewRouteArgs {
     if (identical(this, other)) return true;
     if (other is! OfferFoodOverviewRouteArgs) return false;
     return key == other.key &&
-        const _i31.ListEquality().equals(
+        const _i32.ListEquality().equals(
           initialFoodInfos,
           other.initialFoodInfos,
         );
@@ -936,49 +988,49 @@ class OfferFoodOverviewRouteArgs {
 
   @override
   int get hashCode =>
-      key.hashCode ^ const _i31.ListEquality().hash(initialFoodInfos);
+      key.hashCode ^ const _i32.ListEquality().hash(initialFoodInfos);
 }
 
 /// generated route for
-/// [_i23.OrderShippingOfBoxesScreen]
-class OrderShippingOfBoxesRoute extends _i27.PageRouteInfo<void> {
-  const OrderShippingOfBoxesRoute({List<_i27.PageRouteInfo>? children})
+/// [_i24.OrderShippingOfBoxesScreen]
+class OrderShippingOfBoxesRoute extends _i28.PageRouteInfo<void> {
+  const OrderShippingOfBoxesRoute({List<_i28.PageRouteInfo>? children})
       : super(OrderShippingOfBoxesRoute.name, initialChildren: children);
 
   static const String name = 'OrderShippingOfBoxesRoute';
 
-  static _i27.PageInfo page = _i27.PageInfo(
+  static _i28.PageInfo page = _i28.PageInfo(
     name,
     builder: (data) {
-      return const _i23.OrderShippingOfBoxesScreen();
+      return const _i24.OrderShippingOfBoxesScreen();
     },
   );
 }
 
 /// generated route for
-/// [_i24.ProfileScreen]
-class ProfileRoute extends _i27.PageRouteInfo<void> {
-  const ProfileRoute({List<_i27.PageRouteInfo>? children})
+/// [_i25.ProfileScreen]
+class ProfileRoute extends _i28.PageRouteInfo<void> {
+  const ProfileRoute({List<_i28.PageRouteInfo>? children})
       : super(ProfileRoute.name, initialChildren: children);
 
   static const String name = 'ProfileRoute';
 
-  static _i27.PageInfo page = _i27.PageInfo(
+  static _i28.PageInfo page = _i28.PageInfo(
     name,
     builder: (data) {
-      return const _i24.ProfileScreen();
+      return const _i25.ProfileScreen();
     },
   );
 }
 
 /// generated route for
-/// [_i25.ThankYouScreen]
-class ThankYouRoute extends _i27.PageRouteInfo<ThankYouRouteArgs> {
+/// [_i26.ThankYouScreen]
+class ThankYouRoute extends _i28.PageRouteInfo<ThankYouRouteArgs> {
   ThankYouRoute({
-    _i28.Key? key,
+    _i29.Key? key,
     required bool isSuccess,
     required String message,
-    List<_i27.PageRouteInfo>? children,
+    List<_i28.PageRouteInfo>? children,
   }) : super(
           ThankYouRoute.name,
           args: ThankYouRouteArgs(
@@ -991,11 +1043,11 @@ class ThankYouRoute extends _i27.PageRouteInfo<ThankYouRouteArgs> {
 
   static const String name = 'ThankYouRoute';
 
-  static _i27.PageInfo page = _i27.PageInfo(
+  static _i28.PageInfo page = _i28.PageInfo(
     name,
     builder: (data) {
       final args = data.argsAs<ThankYouRouteArgs>();
-      return _i25.ThankYouScreen(
+      return _i26.ThankYouScreen(
         key: args.key,
         isSuccess: args.isSuccess,
         message: args.message,
@@ -1011,7 +1063,7 @@ class ThankYouRouteArgs {
     required this.message,
   });
 
-  final _i28.Key? key;
+  final _i29.Key? key;
 
   final bool isSuccess;
 
@@ -1036,17 +1088,17 @@ class ThankYouRouteArgs {
 }
 
 /// generated route for
-/// [_i26.WhatsNewScreen]
-class WhatsNewRoute extends _i27.PageRouteInfo<void> {
-  const WhatsNewRoute({List<_i27.PageRouteInfo>? children})
+/// [_i27.WhatsNewScreen]
+class WhatsNewRoute extends _i28.PageRouteInfo<void> {
+  const WhatsNewRoute({List<_i28.PageRouteInfo>? children})
       : super(WhatsNewRoute.name, initialChildren: children);
 
   static const String name = 'WhatsNewRoute';
 
-  static _i27.PageInfo page = _i27.PageInfo(
+  static _i28.PageInfo page = _i28.PageInfo(
     name,
     builder: (data) {
-      return const _i26.WhatsNewScreen();
+      return const _i27.WhatsNewScreen();
     },
   );
 }